# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DoctorStatus = typeof DoctorStatus[keyof typeof DoctorStatus];


export const DoctorStatus = {
  on_time: 'on_time',
  delayed: 'delayed',
  offline: 'offline',
} as const;

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  room: string;
  status: DoctorStatus;
  delayMinutes: number;
  currentToken: number;
  waitingCount: number;
  avgVisitMinutes: number;
}

export interface DoctorDelayInput {
  /**
     * @minimum 0
     * @maximum 240
     */
  delayMinutes: number;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  booked: 'booked',
  checked_in: 'checked_in',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  patientId: number;
  doctorId: number;
  date: string;
  time: string;
  reason: string;
  status: AppointmentStatus;
  /** @nullable */
  tokenNumber?: number | null;
}

export interface Dashboard {
  date: string;
  totalAppointments: number;
  checkedIn: number;
  waitingCount: number;
  currentToken: number;
  avgVisitMinutes: number;
  delayedDoctors: Doctor[];
  nextAppointment: Appointment | null;
}

export type QueueItemSource = typeof QueueItemSource[keyof typeof QueueItemSource];


export const QueueItemSource = {
  online: 'online',
  walk_in: 'walk_in',
} as const;

export type QueueItemStatus = typeof QueueItemStatus[keyof typeof QueueItemStatus];


export const QueueItemStatus = {
  waiting: 'waiting',
  in_room: 'in_room',
  completed: 'completed',
  skipped: 'skipped',
} as const;

export interface QueueItem {
  id: number;
  tokenNumber: number;
  patientId: number;
  patientName: string;
  doctorId: number;
  appointmentTime: string;
  source?: QueueItemSource;
  status: QueueItemStatus;
  estimatedWaitMinutes: number;
}

export type QueueStatusInputStatus = typeof QueueStatusInputStatus[keyof typeof QueueStatusInputStatus];


export const QueueStatusInputStatus = {
  waiting: 'waiting',
  in_room: 'in_room',
  completed: 'completed',
  skipped: 'skipped',
} as const;

export interface QueueStatusInput {
  status: QueueStatusInputStatus;
}

export type TokenInputSource = typeof TokenInputSource[keyof typeof TokenInputSource];


export const TokenInputSource = {
  online: 'online',
  walk_in: 'walk_in',
} as const;

export interface TokenInput {
  patientId?: number;
  doctorId: number;
  appointmentTime?: string;
  source?: TokenInputSource;
  patientName?: string;
  patientPhone?: string;
}

export interface AppointmentInput {
  patientId: number;
  doctorId: number;
  date: string;
  time: string;
  /** @minLength 2 */
  reason: string;
}

export interface Patient {
  id: number;
  name: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  bloodGroup: string;
}

export interface PatientUpdate {
  /** @minLength 2 */
  name?: string;
  email?: string;
  phone?: string;
  dateOfBirth?: string;
  bloodGroup?: string;
}

export interface PrescriptionMedication {
  name: string;
  dosage: string;
  frequency: string;
  duration: string;
}

export interface MedicalDocument {
  name: string;
  objectPath: string;
  contentType: string;
  size: number;
}

export interface Visit {
  id: number;
  patientId: number;
  doctorId: number;
  date: string;
  doctorName: string;
  diagnosis: string;
  notes: string;
  prescriptions: PrescriptionMedication[];
  documents: MedicalDocument[];
}

export interface VisitInput {
  doctorId: number;
  date: string;
  diagnosis: string;
  notes: string;
  prescriptions: PrescriptionMedication[];
  documents?: MedicalDocument[];
}

export interface RequestUploadUrlInput {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResultMetadata = {
  name: string;
  size: number;
  contentType: string;
};

export interface RequestUploadUrlResult {
  uploadURL: string;
  objectPath: string;
  metadata: RequestUploadUrlResultMetadata;
}

export type GetDashboardParams = {
date?: string;
doctorId?: number;
};

export type ListQueueParams = {
doctorId?: number;
date?: string;
};

export type ListAppointmentsParams = {
patientId?: number;
date?: string;
};