# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';
import type { Doctor } from './doctor';

export interface Dashboard {
  date: Date;
  totalAppointments: number;
  checkedIn: number;
  waitingCount: number;
  currentToken: number;
  avgVisitMinutes: number;
  delayedDoctors: Doctor[];
  nextAppointment: Appointment | null;
}