# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DoctorStatus } from './doctorStatus';

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  room: string;
  status: DoctorStatus;
  delayMinutes: number;
  currentToken: number;
  waitingCount: number;
  avgVisitMinutes: number;
}